+++ autoware_utils/README.md
@@ -3,11 +3,3 @@
 ## Overview
 
 The **autoware_utils** library is a comprehensive toolkit designed to facilitate the development of autonomous driving applications. This library provides essential utilities for geometry, mathematics, ROS (Robot Operating System) expansions, diagnostics, and more. It is extensively used in the Autoware project to handle common tasks such as geometric calculations, data normalization, message conversions, performance monitoring, and point cloud transformations.
-
-### Design
-
-#### ROS Module
-
-The ROS module provides utilities for working with ROS messages and nodes:
-
-- **`self_pose_listener.hpp`**: Listens to the self-pose of the vehicle.
